requirements.txt
ultralytics>=8.0.0
opencv-python
python-osc
pillow
